Started over to have better code comprehension
The code before was good, but error-prone because only the  file was imported, and it needed to import files which were importing that same file. In the end, I decided to start over in order to start fresh and have a solid basis. The current files have only been tested with what is actually in , but unit testing should not take long before being added inside the project. For a project as big as this one, I can't make silly mistakes like not testing the code...