Changed types definitions & updated error handling
I realized that with the types before, to store an integer (or a boolean), one would need a struct with an enum (4 bytes) and a pointer (8 bytes) to a struct with the integer (4 bytes), which means a whole 12 bytes lost each time. Because it is mandatory for the string to be a pointer (8 bytes), then every other type below 8 bytes (which means all other types) can be inlined. Furthermore, I racked my brain and did not find any information which may useful to store for these other types. So the type definitions changed. Then, I also changed the way to handle errors. Before, with the ReturnType, there was a major problem, think of the fetch_int function: you fetch an integer, which is a ReturnType whom you take the ret which is a ValueType... so you need to fetch_int it! And so on... So now, whenever a function is called, one needs to check the errno and set it back to 0 again after.